+++ include/cpp_common/combinations.hpp
@@ -60,7 +60,7 @@ get_combinations(const std::vector<II_t_rt>&);
 
 
 std::map<int64_t, std::set<int64_t>>
-get_combinations(char*, ArrayType*, ArrayType*, bool);
+get_combinations(const char*, ArrayType*, ArrayType*, bool);
 
 }  // namespace utilities
 }  // namespace pgrouting