+++ include/drivers/components/biconnectedComponents_driver.h
@@ -48,7 +48,7 @@ extern "C" {
 #endif
 
 void pgr_do_biconnectedComponents(
-        char*,
+        const char*,
 
         II_t_rt**, size_t*,
         char**, char**, char**);