+++ src/bellman_ford/bellman_ford_neg_driver.cpp
@@ -65,9 +65,9 @@ bellman_ford(
 
 void
 pgr_do_bellman_ford_neg(
-                char *edges_sql,
-                char *neg_edges_sql,
-                char *combinations_sql,
+                const char *edges_sql,
+                const char *neg_edges_sql,
+                const char *combinations_sql,
                 ArrayType *starts,
                 ArrayType *ends,
 
@@ -89,7 +89,7 @@ pgr_do_bellman_ford_neg(
     std::ostringstream log;
     std::ostringstream err;
     std::ostringstream notice;
-    char *hint = nullptr;
+    const char *hint = nullptr;
 
     try {
         pgassert(!(*log_msg));