+++ src/components/makeConnected_driver.cpp
@@ -42,7 +42,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
 void
 pgr_do_makeConnected(
-        char *edges_sql,
+        const char *edges_sql,
 
 
                 II_t_rt **return_tuples,
@@ -58,7 +58,7 @@ pgr_do_makeConnected(
     std::ostringstream log;
     std::ostringstream err;
     std::ostringstream notice;
-    char *hint = nullptr;
+    const char *hint = nullptr;
 
     try {
         pgassert(!(*log_msg));