+++ src/pickDeliver/pickDeliverEuclidean_driver.cpp
@@ -47,8 +47,8 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
 void
 pgr_do_pickDeliverEuclidean(
-        char *customers_sql,
-        char *vehicles_sql,
+        const char *customers_sql,
+        const char *vehicles_sql,
 
         double factor,
         int max_cycles,
@@ -67,7 +67,7 @@ pgr_do_pickDeliverEuclidean(
     std::ostringstream log;
     std::ostringstream notice;
     std::ostringstream err;
-    char* hint = nullptr;
+    const char *hint = nullptr;
 
     try {
         *return_tuples = nullptr;