+++ src/spanningTree/prim_driver.cpp
@@ -45,10 +45,10 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
 void
 pgr_do_prim(
-        char *edges_sql,
+        const char *edges_sql,
         ArrayType* starts,
 
-        char* fn_suffix,
+        const char *fn_suffix,
 
         int64_t max_depth,
         double distance,
@@ -67,7 +67,7 @@ pgr_do_prim(
     std::ostringstream log;
     std::ostringstream err;
     std::ostringstream notice;
-    char *hint = nullptr;
+    const char *hint = nullptr;
 
     try {
         // NOLINTBEGIN(clang-analyzer-cplusplus.NewDelete)