+++ doc/src/release_notes.rst
@@ -165,6 +165,8 @@ milestone for 3.8.0
 
 .. rubric:: Deprecation of functions.
 
+* `#2749 <https://github.com/pgRouting/pgrouting/issues/2749>`__:
+  pgr_alphaShape
 * `#2754 <https://github.com/pgRouting/pgrouting/issues/2754>`__:
   pgr_analyzeOneWay
 * `#2753 <https://github.com/pgRouting/pgrouting/issues/2753>`__: