+++ doc/src/release_notes.rst
@@ -96,6 +96,8 @@ milestone for 4.0.0
   pgr_trspviavertices
 * `#2752 <https://github.com/pgRouting/pgrouting/issues/2752>`__:
   pgr_analyzeGraph
+* `#2755 <https://github.com/pgRouting/pgrouting/issues/2755>`__:
+  pgr_analyzeOneWay
 
 .. rubric:: Removal of SQL deprecated internal functions