+++ doc/src/release_notes.rst
@@ -203,7 +203,7 @@ milestone for 3.8.0
      :start-after: Version 3.8.0
      :end-before: .. rubric
 
-.. rubric:: C/C++ code enhancement
+.. rubric:: C/C++ code enhancements
 
 * `#2802 <https://github.com/pgRouting/pgrouting/issues/2802>`__:
   Code factorization on pgr_contraction family