+++ doc/src/release_notes.rst
@@ -104,7 +104,7 @@ milestone for 3.8.0
   * `#2848 <https://github.com/pgRouting/pgrouting/issues/2848>`__:
     pgr_separateCrossing
   * `#2849 <https://github.com/pgRouting/pgrouting/issues/2849>`__:
-    pgr_separateCrossing
+    pgr_separateTouching
 
 pgRouting 3.7
 +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++