+++ doc/src/release_notes.rst
@@ -94,12 +94,14 @@ milestone for 4.0.0
   pgr_trspviaedges
 * `#2682 <https://github.com/pgRouting/pgrouting/issues/2682>`__:
   pgr_trspviavertices
+* `#2748 <https://github.com/pgRouting/pgrouting/issues/2748>`__:
+  pgr_alphaShape
+* `#2751 <https://github.com/pgRouting/pgrouting/issues/2851>`__:
+  pgr_createTopology
 * `#2752 <https://github.com/pgRouting/pgrouting/issues/2752>`__:
   pgr_analyzeGraph
 * `#2755 <https://github.com/pgRouting/pgrouting/issues/2755>`__:
   pgr_analyzeOneWay
-* `#2751 <https://github.com/pgRouting/pgrouting/issues/2851>`__:
-  pgr_createTopology
 * `#2827 <https://github.com/pgRouting/pgrouting/issues/2827>`__:
   pgr_createVerticesTable