+++ doc/src/release_notes.rst
@@ -55,6 +55,7 @@ milestone for 4.0.0
 * pgr_withPointsDD
 * pgr_withPointsKSP
 * pgr_withPointsVia
+* pgr_contractionHierarchies
 
 .. rubric:: Signatures promoted to official
 
@@ -245,6 +246,7 @@ milestone for 3.8.0
 
   * `#2790 <https://github.com/pgRouting/pgrouting/issues/2790>`__: pgr_contractionDeadEnd
   * `#2791 <https://github.com/pgRouting/pgrouting/issues/2791>`__: pgr_contractionLinear
+  * `#2536 <https://github.com/pgRouting/pgrouting/issues/2536>`__: pgr_contractionHierarchies
 
 * Utilities